labs/cloud9/requirements.txt
aws-cdk-lib==2.87.0
constructs>=10.0.0,<11.0.0
aws_cdk.aws_cloud9_alpha
